// frontend/src/api/generated/models/inbound-slip-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipStatus } from './inbound-slip-status';
// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipType } from './inbound-slip-type';

/**
 * 
 * @export
 * @interface InboundSlipSummary
 */
export interface InboundSlipSummary {
    /**
     * 入荷伝票ID
     * @type {number}
     * @memberof InboundSlipSummary
     */
    'id': number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'slipNumber': string;
    /**
     * 
     * @type {InboundSlipType}
     * @memberof InboundSlipSummary
     */
    'slipType': InboundSlipType;
    /**
     * 倉庫コード
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'warehouseCode': string;
    /**
     * 仕入先コード（倉庫振替の場合はnull）
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'partnerCode'?: string | null;
    /**
     * 仕入先名（倉庫振替の場合はnull）
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'partnerName'?: string | null;
    /**
     * 入荷予定日
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'plannedDate': string;
    /**
     * 
     * @type {InboundSlipStatus}
     * @memberof InboundSlipSummary
     */
    'status': InboundSlipStatus;
    /**
     * 明細件数
     * @type {number}
     * @memberof InboundSlipSummary
     */
    'lineCount': number;
    /**
     * 登録日時
     * @type {string}
     * @memberof InboundSlipSummary
     */
    'createdAt': string;
}