// frontend/src/api/generated/models/inbound-slip-summary-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InboundSlipSummary } from './inbound-slip-summary';

/**
 * 
 * @export
 * @interface InboundSlipSummaryPageResponse
 */
export interface InboundSlipSummaryPageResponse {
    /**
     * 
     * @type {Array<InboundSlipSummary>}
     * @memberof InboundSlipSummaryPageResponse
     */
    'content': Array<InboundSlipSummary>;
    /**
     * 
     * @type {number}
     * @memberof InboundSlipSummaryPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof InboundSlipSummaryPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof InboundSlipSummaryPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof InboundSlipSummaryPageResponse
     */
    'totalPages': number;
}